Moving the stream guardian to the common folder
