+++ zakirov.marat/T3/line_commands.cpp
@@ -4,7 +4,7 @@
 #include <limits>
 #include <functional>
 #include <iomanip>
-#include "zakirov.marat/T2/stream_guardian.hpp"
+#include "stream_guardian.hpp"
 
 namespace
 {
