--- zakirov.marat/T3/text.txt
@@ -1,1 +1,0 @@
-
