+++ zakirov.marat/common/stream_guardian.cpp
