+++ zakirov.marat/common/stream_guardian.hpp
